// WYNKWeb/src/app/Models/SurgerySafetyCheckList.ts
export class SurgerySafetyCheckList {
  SSCID: number;
  CmpID: number;
  Question: string;
  Questiontowhom: string;
  DefaultValue: number;
  SSCGroupDescription: number;
  IsDeleted: boolean;
  IsActive: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
}